refactor: extract memory automation + prompt presets from ChatView into composables
- useMemoryAutomation.js: runMemoryAutomationAfterStableTurn, generateMemoryDraftForMessages,
  createPendingMemoryDraft, bootstrapImportedMemoryDrafts, and helper functions
- useAutoSync.js: triggerAutoSyncCheck
- memoryPromptPresets.js: builtInMemoryPrompts, getMemoryPromptOptions, resolveMemoryPrompt, etc.
- ~450 lines removed from ChatView.vue
